refactor(tauri): split asset service modules
